Resolve -Wsign-compare warnings using signed indices
Standardized on signed loop counters (consistent with the existing
index arithmetic like i = x - 1 and j -= 1, which unsigned types would
make underflow-prone). Cache container sizes as const int and compare,
cast size() to int where the size changes mid-loop (dbscan neighbor
expansion), make n an int in main (get_chrom_num already returns int),
and use size_t for the find()/npos comparison in get_linked_clusters.
No behavior change: unit tests, small e2e, and the real-data md5 guards
(transcript-id GTF, assignment GTF, and per-gene counts) all match.

+++ src/impaqt.cpp
@@ -59,7 +59,7 @@ int run_impaqt(int argc, char const ** argv) {
 
    
     // Multithreading Initialization
-    size_t n = processes[init_thread] -> get_chrom_num();
+    int n = processes[init_thread] -> get_chrom_num();
     if (n > 1) {
         processes.reserve(n);
         for (int i = 1; i < n; i++) { processes.emplace_back(std::make_unique<Impaqt>(i)); }
